greengrass/components/cmpe281.edge_resizer/artifacts/requirements.txt
Pillow>=10.0.0
awsiotsdk>=1.21.0
